Brings in 18 new commits from main, including several large
refactors:

- 13a1da6b [BREAKING] MAINT: enforce _async suffix on async functions
- f43855be MAINT: Making Message a Pydantic model
- b3b018ff MAINT: Making Score a Pydantic model
- 414ac8c1 MAINT: Migrate os.path.* to pathlib.Path in pyrit/
- d6173f18 MAINT: Eliminate blocking I/O on async code paths
- 9bb005f1 TEST: stop GCG unit tests from hitting HuggingFace
- 17 conflicts, all resolved by `git checkout --theirs` then
  re-running `ruff check --fix` to apply the PEP 604 sweep to
  main's new code (~149 violations auto-fixed in newly-merged
  files).

No manual hand-fixes needed this time - the runtime
`Optional[dict]` in `pyrit/models/message_piece.py` that
needed hand-fixing in prior merges is gone now that Message and
MessagePiece are full Pydantic models on main.

Verification:
- ruff check pyrit/ tests/ doc/ - clean
- ruff format --check - clean
- pytest tests/unit -n 4 - 9273 passed, 6 skipped, 0 failures

+++ pyrit/auxiliary_attacks/gcg/attack/base/attack_manager.py
@@ -654,7 +654,7 @@ class MultiPromptAttack:
         n_steps: int = 100,
         batch_size: int = 1024,
         topk: int = 256,
-        temp: int = 1,
+        temp: float = 1.0,
         allow_non_ascii: bool = True,
         target_weight: float | None = None,
         control_weight: float | None = None,
